src/is-calibration-server/pyproject.toml
[tool.poetry]
name = "is-calibration-server"
version = "0.2.0"
description = ""
authors = ["Luiz Carlos Cosmi Filho <luizcarloscosmifilho@gmail.com>"]
readme = "README.md"
packages = [{include = "is_calibration_server"}]

[tool.poetry.dependencies]
python = "^3.10"
is-msgs = "1.1.10"
is-wire = "1.2.0"
vine = "1.3.0"
numpy = "1.22.4"
opencv-contrib-python = "4.7.0.68"
opencensus-ext-zipkin = "0.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[tool.poetry.scripts]
is-calibration-server = 'is_calibration_server.service:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
